# api/openapi/shift.yaml
openapi: 3.1.0
info:
  title: SHIFT Control Plane API
  version: 0.1.0
  description: |
    Fleet metadata and orchestration API for SHIFT. Raw computational state,
    checkpoint chunks, and workload keys remain on mutually authenticated agents.
servers:
  - url: http://127.0.0.1:8090
    description: Local development server
...
paths:
...
  /v1/auth/register:
    post:
      tags: [Authentication]
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Account and initial organization created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterResponse'
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { description: The email's domain requires single sign-on; accounts for it arrive through the issuer }
        '409': { $ref: '#/components/responses/Conflict' }
  /v1/auth/login:
    post:
      tags: [Authentication]
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoginRequest' }
      responses:
        '200':
          description: Session tokens
          content:
            application/json:
              schema: { $ref: '#/components/schemas/LoginResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { description: The email's domain requires single sign-on; use the SSO flow instead }
  /v1/auth/refresh:
    post:
      tags: [Authentication]
      operationId: refresh
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RefreshRequest' }
      responses:
        '200':
          description: Rotated session tokens
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SessionTokens' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /v1/auth/sso/authorize:
    post:
      tags: [Authentication]
      operationId: ssoAuthorize
      description: Starts an OIDC single sign-on login. The redirect URI must be a loopback http URL; PKCE S256 is mandatory. Returns the issuer's authorization URL plus the state the client must echo back.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SSOAuthorizeRequest' }
      responses:
        '200':
          description: Authorization URL and state
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SSOAuthorizeResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '502': { description: The issuer could not be reached }
        '503': { description: Single sign-on is not configured }
  /v1/auth/sso/token:
    post:
      tags: [Authentication]
      operationId: ssoToken
      description: Redeems the authorization code the loopback listener received. Validates the ID token's signature, issuer, audience, expiry, and nonce; links or provisions the local account; issues a session.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SSOTokenRequest' }
      responses:
        '200':
          description: Session tokens for the federated account
          content:
            application/json:
              schema: { $ref: '#/components/schemas/LoginResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { description: The ID token failed validation }
        '403': { description: The account is disabled }
        '409': { description: The account is already federated to a different identity }
        '502': { description: The code exchange failed }
        '503': { description: Single sign-on is not configured }
  /v1/auth/sso/required:
    get:
      tags: [Authentication]
      operationId: ssoRequired
      description: Reports whether an email's domain is under SSO enforcement, so clients route to the issuer before asking for a password.
      parameters:
        - name: email
          in: query
          required: true
          schema: { type: string, format: email }
      responses:
        '200':
          description: Enforcement answer
          content:
            application/json:
              schema:
                type: object
                required: [required]
                properties:
                  required: { type: boolean }
        '400': { $ref: '#/components/responses/BadRequest' }
  /v1/auth/logout:
    post:
      tags: [Authentication]
      security: [{ bearerAuth: [] }]
      operationId: logout
      responses:
        '204': { description: Session revoked }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /v1/me:
    get:
      tags: [Authentication]
      security: [{ bearerAuth: [] }]
      operationId: me
      responses:
        '200':
          description: Authenticated principal
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Principal' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /v1/organizations:
    get:
      tags: [Organizations]
      security: [{ bearerAuth: [] }]
      operationId: listOrganizations
      responses:
        '200':
          description: Organizations visible to the user
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Organization' }
  /v1/organizations/{organizationID}/members:
    parameters: [{ $ref: '#/components/parameters/OrganizationID' }]
    post:
      tags: [Organizations]
      security: [{ bearerAuth: [] }]
      operationId: upsertMember
      description: Requires admin or owner role. Owner cannot be assigned by this endpoint.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, role]
              properties:
                email: { type: string, format: email }
                role: { $ref: '#/components/schemas/Role' }
      responses:
        '200': { description: Membership updated }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /v1/organizations/{organizationID}/machines:
    parameters: [{ $ref: '#/components/parameters/OrganizationID' }]
    get:
      tags: [Machines]
      security: [{ bearerAuth: [] }]
      operationId: listMachines
      responses:
        '200':
          description: Registered machines
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Machine' }
    post:
      tags: [Machines]
      security: [{ bearerAuth: [] }]
      operationId: createMachine
      description: Requires operator, admin, or owner role. The agent URL must be HTTPS.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateMachineRequest' }
      responses:
        '201':
          description: Machine registered
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Machine' }
        '402': { $ref: '#/components/responses/EntitlementExceeded' }
        '409': { $ref: '#/components/responses/Conflict' }
  /v1/organizations/{organizationID}/machines/{machineID}/heartbeat:
    parameters:
      - $ref: '#/components/parameters/OrganizationID'
      - $ref: '#/components/parameters/MachineID'
    post:
      tags: [Machines]
      security: [{ bearerAuth: [] }]
      operationId: machineHeartbeat
      description: Updates inventory and online state. Requires operator role.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                agent_url: { type: string, format: uri }
                capabilities: { type: object, additionalProperties: true }
                status: { type: string, enum: [online, offline, draining] }
      responses:
        '200':
          description: Updated machine
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Machine' }
  /v1/organizations/{organizationID}/workloads:
    parameters: [{ $ref: '#/components/parameters/OrganizationID' }]
    get:
      tags: [Workloads]
      security: [{ bearerAuth: [] }]
      operationId: listWorkloads
      responses:
        '200':
          description: Workload metadata
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Workload' }
    post:
      tags: [Workloads]
      security: [{ bearerAuth: [] }]
      operationId: createWorkload
      description: Stores metadata only; execution remains on the selected agent.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateWorkloadRequest' }
      responses:
        '201':
          description: Workload metadata created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Workload' }
  /v1/organizations/{organizationID}/workloads/{workloadID}/status:
    parameters:
      - $ref: '#/components/parameters/OrganizationID'
      - $ref: '#/components/parameters/WorkloadID'
    post:
      tags: [Workloads]
      security: [{ bearerAuth: [] }]
      operationId: updateWorkloadStatus
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, additionalProperties: true }
      responses:
        '200':
          description: Updated workload
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Workload' }
  /v1/organizations/{organizationID}/migrations:
    parameters: [{ $ref: '#/components/parameters/OrganizationID' }]
    get:
      tags: [Migrations]
      security: [{ bearerAuth: [] }]
      operationId: listMigrations
      responses:
        '200':
          description: Migration jobs
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/MigrationJob' }
    post:
      tags: [Migrations]
      security: [{ bearerAuth: [] }]
      operationId: createMigration
      description: Creates a metadata job. Agents execute the actual state transfer over their mutually authenticated peer channel.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateMigrationRequest' }
      responses:
        '202':
          description: Migration queued
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MigrationJob' }
  /v1/organizations/{organizationID}/audit:
    parameters: [{ $ref: '#/components/parameters/OrganizationID' }]
    get:
      tags: [Organizations]
      security: [{ bearerAuth: [] }]
      operationId: listAuditEvents
      description: Requires admin or owner role.
      parameters:
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 500, default: 100 }
      responses:
        '200':
          description: Audit events
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/AuditEvent' }
  /v1/organizations/{organizationID}/machines/capability:
    parameters: [{ $ref: '#/components/parameters/OrganizationID' }]
    get:
      tags: [Machines]
      security: [{ bearerAuth: [] }]
      operationId: queryMachinesByCapability
      description: >
        Fleet query answered from the relational projection of each machine's
        JSONB capability document. Only scalar capabilities (number, text,
        boolean) are projected; objects and arrays are not queryable.
      parameters:
        - name: capability
          in: query
          required: true
          schema: { type: string, maxLength: 256 }
        - name: kind
          in: query
          required: true
          schema: { type: string, enum: [number, text, boolean] }
        - name: minimum
          in: query
          required: false
          description: Value floor; valid only with kind=number.
          schema: { type: number, format: double }
      responses:
        '200':
          description: Machines reporting the capability, numbers ordered by value descending
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Machine' }
        '400':
          description: Missing capability, unknown kind, or minimum without kind=number
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /v1/organizations/{organizationID}/retention:
    parameters: [{ $ref: '#/components/parameters/OrganizationID' }]
    get:
      tags: [Organizations]
      security: [{ bearerAuth: [] }]
      operationId: getRetentionPolicy
      description: Requires admin or owner role. Returns defaults when no policy was configured.
      responses:
        '200':
          description: Retention windows in days
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RetentionPolicy' }
    put:
      tags: [Organizations]
      security: [{ bearerAuth: [] }]
      operationId: setRetentionPolicy
      description: >
        Requires admin or owner role. Omitted fields keep their current value;
        zero is the deliberate keep-forever policy. Changes are audited and
        take effect on the next enforcement sweep.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SetRetentionPolicyRequest' }
      responses:
        '200':
          description: Stored policy
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RetentionPolicy' }
        '400':
          description: Unknown field or a window outside 0-36500 days
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /v1/organizations/{organizationID}/sso:
    parameters: [{ $ref: '#/components/parameters/OrganizationID' }]
    get:
      tags: [Organizations]
      security: [{ bearerAuth: [] }]
      operationId: getOrganizationSSO
      description: Requires admin or owner role. Reports the organization's SSO enforcement state.
      responses:
        '200':
          description: Enforcement state
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OrganizationSSO' }
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      tags: [Organizations]
      security: [{ bearerAuth: [] }]
      operationId: setOrganizationSSO
      description: >
        Requires admin or owner role. Enforcing SSO claims an email domain:
        from that moment the domain's accounts authenticate through the
        configured OIDC issuer and local passwords stop working for them. A
        domain can be claimed by one organization at a time. Disabling clears
        the claim. Requires the control plane to have an OIDC issuer
        configured.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/OrganizationSSO' }
      responses:
        '200':
          description: Stored enforcement state
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OrganizationSSO' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '409': { description: Another organization already claims this domain }
        '503': { description: Single sign-on is not configured }
  /v1/organizations/{organizationID}/entitlement:
    parameters: [{ $ref: '#/components/parameters/OrganizationID' }]
    get:
      tags: [Billing]
      security: [{ bearerAuth: [] }]
      operationId: getEntitlement
      responses:
        '200':
          description: Current server-side limits
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Entitlement' }
  /v1/organizations/{organizationID}/storage:
    parameters: [{ $ref: '#/components/parameters/OrganizationID' }]
    get:
      tags: [Storage]
      security: [{ bearerAuth: [] }]
      operationId: getStorageStatus
      description: |
        Platform-hosted checkpoint storage status. Usage is metered by
        reconciling actual object bytes in the organization's bucket prefix —
        never from agent claims — and refreshes synchronously when the stored
        figures are stale. When hosting is disabled the response is
        `{enabled: false}` with no location fields.
      responses:
        '200':
          description: Storage status
          content:
            application/json:
              schema: { $ref: '#/components/schemas/StorageStatus' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /v1/organizations/{organizationID}/storage/credentials:
    parameters: [{ $ref: '#/components/parameters/OrganizationID' }]
    get:
      tags: [Storage]
      security: [{ bearerAuth: [] }]
      operationId: getStorageCredentials
      description: |
        Short-lived, organization-scoped object-store credentials for agent
        mirroring. Agents call this with a machines-scope API key and refresh
        at half the credential lifetime. Over quota the response is 403 and
        no credentials are issued — mirroring stops while local checkpointing
        continues.
      responses:
        '200':
          description: Issued credentials; the secrets are shown once and never logged
          content:
            application/json:
              schema: { $ref: '#/components/schemas/StorageCredentials' }
        '403':
          description: Over plan quota — issuance suspended until usage drops or the plan is upgraded
          content:
            application/json:
              schema:
                type: object
                required: [code, message, used_storage_bytes, max_storage_bytes]
                properties:
                  code: { type: string, enum: [STORAGE_QUOTA_EXCEEDED] }
                  message: { type: string }
                  used_storage_bytes: { type: integer, format: int64 }
                  max_storage_bytes: { type: integer, format: int64 }
        '503':
          description: This control plane does not host checkpoint storage
          content:
            application/json:
              schema:
                type: object
                required: [code, message]
                properties:
                  code: { type: string, enum: [STORAGE_NOT_CONFIGURED] }
                  message: { type: string }
  /v1/organizations/{organizationID}/storage/reconcile:
    parameters: [{ $ref: '#/components/parameters/OrganizationID' }]
    post:
      tags: [Storage]
      security: [{ bearerAuth: [] }]
      operationId: reconcileStorage
      description: Forces a usage recount from the bucket. Admin, billing scope.
      responses:
        '200':
          description: Refreshed status
          content:
            application/json:
              schema: { $ref: '#/components/schemas/StorageStatus' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '503':
          description: This control plane does not host checkpoint storage
          content:
            application/json:
              schema:
                type: object
                required: [code, message]
                properties:
                  code: { type: string, enum: [STORAGE_NOT_CONFIGURED] }
                  message: { type: string }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: SHIFT access token
...
  responses:
    BadRequest:
      description: Invalid request
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
...
    Conflict:
      description: Resource conflict
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
...
  schemas:
...
    Role:
      type: string
      enum: [owner, admin, operator, viewer]
    RegisterRequest:
      type: object
      required: [email, password, display_name]
      properties:
        email: { type: string, format: email }
        password: { type: string, minLength: 12, writeOnly: true }
        display_name: { type: string, maxLength: 200 }
        organization: { type: string, maxLength: 200 }
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string, writeOnly: true }
...
    RegisterResponse:
      type: object
      required: [user, organization, tokens]
      properties:
        user: { $ref: '#/components/schemas/User' }
        organization: { $ref: '#/components/schemas/Organization' }
        tokens: { $ref: '#/components/schemas/SessionTokens' }
    LoginResponse:
      type: object
      required: [user, tokens]
      properties:
        user: { $ref: '#/components/schemas/User' }
        tokens: { $ref: '#/components/schemas/SessionTokens' }
...